requirements.txt
djangorestframework-simplejwt==5.2.0
django-filter
markdown
djangorestframework~=3.13.1
django~=4.0.6
django-environ
psycopg2-binary
django-cors-headers
Pillow
social-auth-app-django==5.0.0
djoser
drf_yasg
